Fix more occurences of du<>Sie (informal/polite) addressing the user
To have one consistent usage of informal or polite adressing a customer
it is needed to change this. These changes change into the more polite
version becuase most of the content was already in this form and this
can be used easier for other German translations like for the rest of
the DACH region if you need more localized i18n for Swiss or Austrian
shops.