Fixed two issues
The first one was that an null value in an state was changed to an object.
The second one was that now an null value was not changed to an object we
need to change the null to an object when we try to change the type the
original state.